implementata funzione di swap ma ancora non funzionante (errore di segmentation fault). Aggiunte delle printf di controllo in insertion_sort e swap_elem

+++ Esercizio_1/src/array_to_sort.h
@@ -49,7 +49,7 @@ void *array_to_sort_get(UnsortedArray *, unsigned long);
 void array_to_sort_free_memory(UnsortedArray *);
 
 
-void array_to_sort_swap_elem(UnsortedArray *, int, int);
+void array_to_sort_swap_elem(UnsortedArray *, unsigned long, unsigned long);
 
 /* It takes an unsorted array as an input and it orders with a hybrid
  * bin-insert-merge-sort algorithm.
